+++ src/services/UserService.ts
@@ -1,26 +1,26 @@
 import axios from "axios";
+import { AxiosResponse } from "axios";
 import { ServerDTO } from "../models/ServerDTO";
-import * as https from 'https'
+import { Agent } from 'https'
 
 export class UserService {
 
     /**
      * Obter informações do usuário
-     * @param server 
-     * @returns 
      */
-    public static async getUser(server: ServerDTO) {
-        let uri = server.ssl ? "https://" : "http://";
-        uri += server.host;
-        uri += ":" + server.port;
-        uri += "/portal/api/rest/wcmservice/rest/user/findUserByLogin";
-        uri += "?username=" + server.username;
-        uri += "&password=" + server.password;
-        uri += "&login=" + server.username;
+    public static async getUser(server: ServerDTO): Promise<AxiosResponse<any>> {
+        const uri: string = (server.ssl ? "https://" : "http://")
+            + server.host
+            + ":" + server.port
+            + "/portal/api/rest/wcmservice/rest/user/findUserByLogin"
+            + "?username=" + server.username
+            + "&password=" + server.password
+            + "&login=" + server.username;
 
-        const agent = new https.Agent({
+        const agent = new Agent({
             rejectUnauthorized: false
-        })
+        });
+
         return await axios.get(uri, {
             httpsAgent: agent
         });